fix(spi,codecs): write the real field number; stored fields are byte-exact
codecs 134 -> 111. The stored-fields write path now produces bytes
identical to Apache Lucene's.

  _0.fdt  LUCENE 602 bytes  sha256=3d110b19...d672dd
  _0.fdt  GOCENE 602 bytes  sha256=3d110b19...d672dd   0 differing positions
  _0.fdx  identical (64)      _0.fdm  identical (157)

spi.StoredFieldsWriter.WriteField took only an IndexableField, which
carries no field number, so the writer emitted the per-document sequential
index where Lucene emits (info.number << TYPE_BITS) | type - 0x1a instead
of 0x5a for field 11. All seven Java writeField overloads take FieldInfo
first (StoredFieldsWriter.java:63-87), so the signature now does too, swept
across 9 implementers and 6 call sites.

Three call sites had no FieldInfo to pass. Two are resolved exactly as
Java resolves them: SortingStoredFieldsConsumer's copyVisitor and
SegmentMerger's merge visitor now look the field up by name against the
same FieldInfos Lucene hands its own visitor - which is literally
StoredFieldsWriter.java:196-203's remapper.fieldInfo(field.name) against
mergeState.mergeFieldInfos. Both fail loudly if the field is absent.

Attribution is proven, not argued: the same harness against a tree pinned
at HEAD fails with "602 bytes, 602 bytes, 22 differing byte positions,
first at offset 170" and the 5a-vs-1a hexdump. One change, opposite
verdict. The regression test committed red last commit is now green, all
three cases.

writeZFloat and writeZDouble wrote their multi-byte tail big-endian where
Java builds it from DataOutput.writeShort/writeInt/writeLong, which are
little-endian. Both directions are now literal transcriptions through
store's own little-endian primitives, and the four Java branches are
proven byte for byte against a Java-computed expectation: 1.5f is
3f 00 c0 00, 1.5d is fe 00 00 c0 3f. The reader was verified independently
by decoding the same seven values back.

StoredFieldsInts.writeInts32 now reproduces Java's sign-extension instead
of masking to 32 bits - immaterial for the high operand after the shift,
but the low one is OR'd unshifted, so a negative value sets all 64 bits.

The three call sites left dangling by the previous commit are resolved
through the port's existing init-time registration seam, extended with two
registrations in the same shape. That was required rather than convenient:
Lucene104Codec needs new Lucene90StoredFieldsFormat(mode.storedMode), and
that format's BEST_SPEED_MODE is itself declared in codecs/lucene90, which
package codecs cannot name.

Also fixed: MergeState.TermVectorsReaders added and populated as
MergeState.java:150 does; three sites iterating mergeFieldInfos corrected
to match Java's direct FieldInfos iteration; the Lucene90DocValuesConsumer
terms-dict buffer calls transcribed exactly - Java genuinely uses
ByteArrayDataOutput there, only the call shapes had drifted; and a
corrupted import line in mutable_point_tree.go where a tab had become a
literal t.

Left red deliberately: SegmentCommitInfo.FieldInfosFiles is a map where
WriteSetOfStrings takes a slice, and Java iterates a HashSet in
String-hash order, so converting forces an ordering choice that changes
.si bytes. Two equally-compiling renderings differing in serialized form
is a Source Fidelity Mandate point 4 case.

+++ codecs/lucene90/compressing/stored_fields_ints.go
@@ -152,15 +152,23 @@ func storedFieldsIntsWriteInts16(out store.DataOutput, count int, values []int32
 }
 
 // storedFieldsIntsWriteInts32 is StoredFieldsInts.writeInts32
-// (StoredFieldsInts.java:97-112).
+// (StoredFieldsInts.java:97-112):
+//
+//	long l = ((long) values[step + i] << 32) | (long) values[step + 64 + i];
+//
+// Both operands are Java int values widened by a SIGN-extending (long) cast.
+// The widening of the high operand is immaterial once it is shifted left by
+// 32, but the low operand is OR'd in unshifted, so a negative value there
+// sets every one of the 64 bits. Masking it to 32 bits would set only the
+// low half and emit different bytes; the port therefore widens through int64,
+// exactly as Java does.
 func storedFieldsIntsWriteInts32(out store.DataOutput, count int, values []int32, offset int) error {
 	k := 0
 	for ; k < count-storedFieldsIntsBlockSizeMinusOne; k += storedFieldsIntsBlockSize {
 		step := offset + k
 		for i := 0; i < 64; i++ {
-			l := uint64(uint32(values[step+i]))<<32 |
-				uint64(uint32(values[step+64+i]))
-			if err := out.WriteLong(int64(l)); err != nil {
+			l := (int64(values[step+i]) << 32) | int64(values[step+64+i])
+			if err := out.WriteLong(l); err != nil {
 				return fmt.Errorf("lucene90/compressing: StoredFieldsInts writeInts32 long: %w", err)
 			}
 		}
